+++ lib/features/environment_summary/application/environment_summary_actions_provider.g.dart
@@ -1,0 +1,27 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of 'environment_summary_actions_provider.dart';
+
+// **************************************************************************
+// RiverpodGenerator
+// **************************************************************************
+
+String _$environmentSummaryActionsHash() =>
+    r'839d87260e34a564590ef79ae7cba9e475b251d7';
+
+/// See also [EnvironmentSummaryActions].
+@ProviderFor(EnvironmentSummaryActions)
+final environmentSummaryActionsProvider =
+    AutoDisposeNotifierProvider<EnvironmentSummaryActions, void>.internal(
+      EnvironmentSummaryActions.new,
+      name: r'environmentSummaryActionsProvider',
+      debugGetCreateSourceHash: const bool.fromEnvironment('dart.vm.product')
+          ? null
+          : _$environmentSummaryActionsHash,
+      dependencies: null,
+      allTransitiveDependencies: null,
+    );
+
+typedef _$EnvironmentSummaryActions = AutoDisposeNotifier<void>;
+// ignore_for_file: type=lint
+// ignore_for_file: subtype_of_sealed_class, invalid_use_of_internal_member, invalid_use_of_visible_for_testing_member, deprecated_member_use_from_same_package